requirements.txt
python-dotenv==1.0.1
pytz==2024.1
Requests==2.31.0
schedule==1.2.1
twitchAPI==4.2.0
twitchAPI==4.2.0
utils==1.0.2
ffpb==0.4.1